+++ realm/realm-jni/src/util.cpp
@@ -160,6 +160,12 @@ void ThrowException(JNIEnv* env, ExceptionKind exception, const std::string& cla
             jExceptionClass = env->FindClass("io/realm/internal/async/BadVersionException");
             message = classStr;
             break;
+
+        case InvalidDb:
+            jExceptionClass = env->FindClass("io/realm/exceptions/RealmInvalidDatabaseException");
+            message = classStr;
+            break;
+
     }
     if (jExceptionClass != NULL) {
         env->ThrowNew(jExceptionClass, message.c_str());