+++ realm/realm-library/src/main/java/io/realm/internal/Util.java
@@ -67,7 +67,8 @@ public class Util {
         Exception_RowInvalid(13),
         Exception_EncryptionNotSupported(14),
         Exception_CrossTableLink(15),
-        Exception_BadVersion(16);
+        Exception_BadVersion(16),
+        Exception_InvalidDatabase(17);
 
         private final int nativeTestcase;
         Testcase(int nativeValue) {