+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+### Breaking changes
+
+* A `RealmInvalidDatabaseException` is thrown instead of `IllegalStateException` when opening a corrupted Realm file, a non-Realm file, a Realm file created by newer version of Realm, or an encrypted Realm file with a wrong key.
+
 ## 1.0.1
 
 ### Bug fixes